Move renderer targets to common interface, to enable using them il all renderers
